+++ libs/cardano-data/test/Test/Cardano/Data/OSet/StrictSpec.hs
@@ -9,7 +9,8 @@ import Cardano.Ledger.Binary (natVersion)
 import Control.Monad (forM_)
 import Data.OSet.Strict hiding (empty)
 import Data.Proxy
-import Data.Sequence.Strict (StrictSeq, fromList, (><))
+import Data.Sequence.Strict (StrictSeq, (><))
+import qualified Data.Sequence.Strict as SSeq (fromList)
 import Data.Set (Set, elems, empty)
 import Test.Cardano.Data.Arbitrary ()
 import Test.Cardano.Ledger.Binary.RoundTrip (cborTrip, embedTripSpec, roundTripCborSpec)
@@ -69,17 +70,17 @@ spec =
         oset ><| oset' `shouldSatisfy` invariantHolds'
         fromStrictSeq sseq `shouldSatisfy` invariantHolds'
         fromSet set `shouldSatisfy` invariantHolds'
-    prop "fromStrictSeq preserves order" $
+    prop "fromList preserves order" $
       \(set :: Set Int) ->
-        let sseq = fromList $ elems set
-         in toStrictSeq (fromStrictSeq sseq) `shouldBe` sseq
+        let sseq = SSeq.fromList $ elems set
+         in toStrictSeq (fromList (elems set)) `shouldBe` sseq
     context "fromStrictSeqDuplicates" $ do
       prop "with duplicates" $ \(set :: (Set Int)) ->
-        let sseq = fromList $ elems set
+        let sseq = SSeq.fromList $ elems set
             oset = fromStrictSeq sseq
          in fromStrictSeqDuplicates (sseq >< sseq) `shouldBe` (set, oset)
       prop "without duplicates" $ \(set :: (Set Int)) ->
-        let sseq = fromList $ elems set
+        let sseq = SSeq.fromList $ elems set
             oset = fromStrictSeq sseq
          in fromStrictSeqDuplicates sseq `shouldBe` (empty, oset)
     context "CBOR round-trip" $ do